# input generation: high depth region speedup