@keiranmraine 2019-05-06
Additional info:

WGS pair reduced wall time from `31:46:27` to `16:37:36` (h:mm:ss)