@AndyMenzies 2019-05-07
Changes all look reasonable to me.  Speed up looks great.